add tests on solutions and renamed by software standards
