+++ tests/test_problem_zoo/test_allen_cahn_problem.py
