+++ tests/test_problem_zoo/test_diffusion_reaction_problem.py
@@ -1,4 +1,5 @@
 import pytest
+import torch
 from pina.problem.zoo import DiffusionReactionProblem
 from pina.problem import TimeDependentProblem, SpatialProblem
 
@@ -17,3 +18,19 @@ def test_constructor(alpha):
     # Should fail if alpha is not a float or int
     with pytest.raises(ValueError):
         problem = DiffusionReactionProblem(alpha="invalid")
+
+
+@pytest.mark.parametrize("alpha", [0.1, 1])
+def test_solution(alpha):
+
+    # Find the solution to the problem
+    problem = DiffusionReactionProblem(alpha=alpha)
+    problem.discretise_domain(n=10, mode="grid", domains=None)
+    pts = problem.discretised_domains["D"]
+    solution = problem.solution(pts.requires_grad_())
+
+    # Compute the residual
+    residual = problem.conditions["D"].equation.residual(pts, solution).tensor
+
+    # Assert the residual of the PDE is close to zero
+    assert torch.allclose(residual, torch.zeros_like(residual), atol=5e-5)
